Cambiado el timer consultado de 8bits (TCNT0) por el de 16bits (TCNT1).

+++ accelerometer.h
@@ -24,8 +24,8 @@ class Accelerometer{
 		Accelerometer();
 		bool getData(Vector<int> &vector);
 
-		uint8_t deltaT;
-		uint8_t deltaT_old;
+		uint16_t deltaT;
+		uint16_t deltaT_old;
 
 	private:
 		bool dataReady();
